delete file parse_args.go
